+++ src/ast/expr/mod.rs
@@ -1,9 +1,11 @@
+mod cast;
 mod expr;
 mod literal;
 mod operator;
 mod syscall;
 mod unary;
 
+pub use cast::*;
 pub use expr::*;
 pub use literal::*;
 pub use operator::*;
